+++ BilleteraVirtualAPI/billeteravirtual/src/main/java/ar/com/ada/api/billeteravirtual/models/response/TransferenciaResponse.java
@@ -1,12 +1,14 @@
 package ar.com.ada.api.billeteravirtual.models.response;
 
+import java.math.BigDecimal;
+
 public class TransferenciaResponse {
 
 
     public boolean isOk = false;
     public String message = "";
-    public int billeteraIdOrigen;
-    public int billeteraIdDestino;
-    public double importe;
-    public int operacionId;
+    public Integer billeteraIdOrigen;
+    public Integer billeteraIdDestino;
+    public BigDecimal importe;
+    public Integer operacionId;
 }
